Admin concurrency

+++ Hooking/Controllers/AdventureAppealsController.cs
@@ -69,8 +69,21 @@ namespace Hooking.Controllers
             string instructorEmail = GetInstructorEmailFromAppeal(appeal);
             await _emailSender.SendEmailAsync(instructorEmail, "Odgovor na žalbu", answer);
             appeal = _context.AdventureAppeal.FirstOrDefault(a => a.Id == appeal.Id);
+            if (appeal == null)
+            {
+                Debug.WriteLine("Concurrency error!");
+                return RedirectToAction("ConcurrencyError", "Home");
+            }
             _context.AdventureAppeal.Remove(appeal);
-            await _context.SaveChangesAsync();
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                Debug.WriteLine("Concurrency error!");
+                return RedirectToAction("ConcurrencyError", "Home");
+            }
 
             return RedirectToAction(nameof(Index));
         }
